Simplify date addition using date-fns
