+++ package.json
@@ -50,6 +50,7 @@
   "dependencies": {
     "aws-sdk": "^2.655.0",
     "aws-xray-sdk-core": "3.2.0",
+    "date-fns": "^2.21.1",
     "date-fns-tz": "^1.1.4",
     "debug": "4.3.1",
     "qrcode": "^1.4.4",
